requirements.txt
hikari
black
hikari-lightbulb
sqlalchemy[sqlite]